requirements.txt
albumentations==1.4.6
annotated-types==0.6.0
appdirs==1.4.4
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
docker-pycreds==0.4.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.14.0
fonttools==4.51.0
fsspec==2024.3.1
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
imageio==2.34.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
install==1.3.5
iopath==0.1.10
ipykernel==6.29.3
ipython==8.18.1
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
lazy_loader==0.4
lightning-utilities==0.11.2
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
matplotlip==0.2
mpmath==1.3.0
nest_asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pip==23.3.1
platformdirs==4.2.1
portalocker==2.8.2
pretty-errors==1.2.25
prompt-toolkit==3.0.42
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
scikit-image==0.22.0
scikit-learn==1.4.2
scipy==1.13.0
sentry-sdk==2.1.1
setproctitle==1.3.3
setuptools==69.5.1
six==1.16.0
smmap==5.0.1
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.5.10
torch==2.3.0
torchmetrics==1.4.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
triton==2.3.0
typing_extensions==4.11.0
urllib3==2.2.1
wandb==0.16.6
wcwidth==0.2.13
wheel==0.43.0
widgetsnbextension==4.0.10
yacs==0.1.8
zipp==3.17.0
